minor changes for less warnings

+++ Library/md5.cpp
@@ -111,7 +111,7 @@ MD5::MD5()
 MD5::MD5(const std::string& text)
 {
     init();
-    update(text.c_str(), text.length());
+    update(text.c_str(), (unsigned int)text.length());
     finalize();
 }
